+++ src/core/db/repository/administrator_repository.py
@@ -5,6 +5,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from src.core.db.db import get_session
 from src.core.db.models import Administrator
 from src.core.db.repository import AbstractRepository
+from src.core.exceptions import AdministratorNotFoundException
 
 
 class AdministratorRepository(AbstractRepository):
@@ -13,11 +14,14 @@ class AdministratorRepository(AbstractRepository):
     def __init__(self, session: AsyncSession = Depends(get_session)) -> None:
         super().__init__(session, Administrator)
 
-    async def get_by_email_or_none(self, email: str) -> Administrator:
-        """Получает из БД администратора по его email. В случае отсутствия возвращает None.
+    async def get_by_email(self, email: str) -> Administrator:
+        """Получает из БД администратора по его email. В случае отсутствия бросает ошибку.
 
         Аргументы:
             email (str) - email администратора.
         """
         administrator = await self._session.execute(select(Administrator).where(Administrator.email == email))
-        return administrator.scalars().first()
+        administrator = administrator.scalars().first()
+        if not administrator:
+            raise AdministratorNotFoundException()
+        return administrator
